requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
psycopg2-binary==2.9.6
pandas==2.0.2
numpy==1.24.4
pytest==7.3.1
python-dotenv==1.0.0
